requirements.txt
falcon==1.4.1
gunicorn==19.9.0
cassandra-driver==3.14.0

# test deps
pytest==3.6.3
requests==2.19.1
pytest-ordering==0.5